feat: refactor home
- discussion and tasks
