+++ frontend/components.d.ts
@@ -47,6 +47,7 @@ declare module 'vue' {
     LucideBellPlus: typeof import('~icons/lucide/bell-plus')['default']
     LucideCalendar: typeof import('~icons/lucide/calendar')['default']
     LucideCheck: typeof import('~icons/lucide/check')['default']
+    LucideChevronRight: typeof import('~icons/lucide/chevron-right')['default']
     LucideEdit: typeof import('~icons/lucide/edit')['default']
     LucideEdit2: typeof import('~icons/lucide/edit2')['default']
     LucideEdit3: typeof import('~icons/lucide/edit3')['default']
@@ -59,13 +60,16 @@ declare module 'vue' {
     LucideRefreshCw: typeof import('~icons/lucide/refresh-cw')['default']
     LucideRotateCcw: typeof import('~icons/lucide/rotate-ccw')['default']
     LucideSave: typeof import('~icons/lucide/save')['default']
+    LucideSearch: typeof import('~icons/lucide/search')['default']
     LucideUnlock: typeof import('~icons/lucide/unlock')['default']
     LucideUserPlus: typeof import('~icons/lucide/user-plus')['default']
     LucideUsers: typeof import('~icons/lucide/users')['default']
     LucideX: typeof import('~icons/lucide/x')['default']
     Members: typeof import('./src/components/Settings/Members.vue')['default']
     MobileLayout: typeof import('./src/components/MobileLayout.vue')['default']
+    NewTaskDialog: typeof import('./src/components/NewTaskDialog.vue')['default']
     PageBreadcrumbs: typeof import('./src/components/PageBreadcrumbs.vue')['default']
+    PageCards: typeof import('./src/components/PageCards.vue')['default']
     PageList: typeof import('./src/components/PageList.vue')['default']
     Pie: typeof import('./src/components/Pie.vue')['default']
     Poll: typeof import('./src/components/Poll.vue')['default']
@@ -81,6 +85,7 @@ declare module 'vue' {
     TabButtons: typeof import('./src/components/TabButtons.vue')['default']
     Tabs: typeof import('./src/components/Tabs.vue')['default']
     TaskCheck: typeof import('./src/components/TaskCheck.vue')['default']
+    TaskDetail: typeof import('./src/components/TaskDetail.vue')['default']
     TaskList: typeof import('./src/components/TaskList.vue')['default']
     TaskPriorityIcon: typeof import('./src/components/icons/TaskPriorityIcon.vue')['default']
     TaskStatusIcon: typeof import('./src/components/icons/TaskStatusIcon.vue')['default']
