+++ frontend/src/pages/HomeTask.vue
@@ -1,0 +1,9 @@
+<template>
+  <TaskDetail :taskId="taskId" />
+</template>
+<script>
+export default {
+  name: 'HomeTask',
+  props: ['taskId'],
+}
+</script>
